Swift: Split session tests into a separate file
