+++ swift/Tests/LinuxMain.swift
@@ -7,6 +7,7 @@ import XCTest
 @testable import SignalClientTests
 
 XCTMain([
-     testCase(ClonableHandleOwnerTests.allTests),
-     testCase(PublicAPITests.allTests),
+    testCase(ClonableHandleOwnerTests.allTests),
+    testCase(PublicAPITests.allTests),
+    testCase(SessionTests.allTests),
 ])
