+++ swift/Tests/SignalClientTests/PublicAPITests.swift
@@ -234,109 +234,6 @@ class PublicAPITests: XCTestCase {
         XCTAssertEqual(b_ptext, [1, 2, 3])
     }
 
-    fileprivate func initializeSessions(alice_store: InMemorySignalProtocolStore,
-                                        bob_store: InMemorySignalProtocolStore,
-                                        bob_address: ProtocolAddress) {
-        let bob_pre_key = try! PrivateKey.generate()
-        let bob_signed_pre_key = try! PrivateKey.generate()
-
-        let bob_signed_pre_key_public = try! bob_signed_pre_key.publicKey().serialize()
-
-        let bob_identity_key = try! bob_store.identityKeyPair(context: nil).identityKey
-        let bob_signed_pre_key_signature = try! bob_store.identityKeyPair(context: nil).privateKey.generateSignature(message: bob_signed_pre_key_public)
-
-        let prekey_id: UInt32 = 4570
-        let signed_prekey_id: UInt32 = 3006
-
-        let bob_bundle = try! PreKeyBundle(registrationId: try! bob_store.localRegistrationId(context: nil),
-                                           deviceId: 9,
-                                           prekeyId: prekey_id,
-                                           prekey: bob_pre_key.publicKey(),
-                                           signedPrekeyId: signed_prekey_id,
-                                           signedPrekey: try! bob_signed_pre_key.publicKey(),
-                                           signedPrekeySignature: bob_signed_pre_key_signature,
-                                           identity: bob_identity_key)
-
-        // Alice processes the bundle:
-        try! processPreKeyBundle(bob_bundle,
-                                 for: bob_address,
-                                 sessionStore: alice_store,
-                                 identityStore: alice_store,
-                                 context: nil)
-
-        XCTAssertEqual(try! alice_store.loadSession(for: bob_address, context: nil)?.remoteRegistrationId(),
-                       try! bob_store.localRegistrationId(context: nil))
-
-        // Bob does the same:
-        try! bob_store.storePreKey(PreKeyRecord(id: prekey_id, privateKey: bob_pre_key),
-                                   id: prekey_id,
-                                   context: nil)
-
-        try! bob_store.storeSignedPreKey(
-            SignedPreKeyRecord(
-                id: signed_prekey_id,
-                timestamp: 42000,
-                privateKey: bob_signed_pre_key,
-                signature: bob_signed_pre_key_signature
-            ),
-            id: signed_prekey_id,
-            context: nil)
-    }
-
-    func testSessionCipher() {
-        let alice_address = try! ProtocolAddress(name: "+14151111111", deviceId: 1)
-        let bob_address = try! ProtocolAddress(name: "+14151111112", deviceId: 1)
-
-        let alice_store = try! InMemorySignalProtocolStore()
-        let bob_store = try! InMemorySignalProtocolStore()
-
-        initializeSessions(alice_store: alice_store, bob_store: bob_store, bob_address: bob_address)
-
-        // Alice sends a message:
-        let ptext_a: [UInt8] = [8, 6, 7, 5, 3, 0, 9]
-
-        let ctext_a = try! signalEncrypt(message: ptext_a,
-                                         for: bob_address,
-                                         sessionStore: alice_store,
-                                         identityStore: alice_store,
-                                         context: nil)
-
-        XCTAssertEqual(try! ctext_a.messageType(), .preKey)
-
-        let ctext_b = try! PreKeySignalMessage(bytes: try! ctext_a.serialize())
-
-        let ptext_b = try! signalDecryptPreKey(message: ctext_b,
-                                               from: alice_address,
-                                               sessionStore: bob_store,
-                                               identityStore: bob_store,
-                                               preKeyStore: bob_store,
-                                               signedPreKeyStore: bob_store,
-                                               context: nil)
-
-        XCTAssertEqual(ptext_a, ptext_b)
-
-        // Bob replies
-        let ptext2_b: [UInt8] = [23]
-
-        let ctext2_b = try! signalEncrypt(message: ptext2_b,
-                                          for: alice_address,
-                                          sessionStore: bob_store,
-                                          identityStore: bob_store,
-                                          context: nil)
-
-        XCTAssertEqual(try! ctext2_b.messageType(), .whisper)
-
-        let ctext2_a = try! SignalMessage(bytes: try! ctext2_b.serialize())
-
-        let ptext2_a = try! signalDecrypt(message: ctext2_a,
-                                          from: bob_address,
-                                          sessionStore: alice_store,
-                                          identityStore: alice_store,
-                                          context: nil)
-
-        XCTAssertEqual(ptext2_a, ptext2_b)
-    }
-
     func testSenderCertificates() {
         let senderCertBits: [UInt8] = [
           0x0a, 0xcd, 0x01, 0x0a, 0x0c, 0x2b, 0x31, 0x34, 0x31, 0x35, 0x32, 0x32, 0x32, 0x32, 0x32, 0x32, 0x32, 0x10, 0x2a, 0x19,
@@ -373,50 +270,6 @@ class PublicAPITests: XCTestCase {
         XCTAssertEqual(try! serverCert.signatureBytes().count, 64)
     }
 
-    func testSealedSenderSession() throws {
-        let alice_address = try! ProtocolAddress(name: "+14151111111", deviceId: 1)
-        let bob_address = try! ProtocolAddress(name: "+14151111112", deviceId: 1)
-
-        let alice_store = try! InMemorySignalProtocolStore()
-        let bob_store = try! InMemorySignalProtocolStore()
-
-        initializeSessions(alice_store: alice_store, bob_store: bob_store, bob_address: bob_address)
-
-        let trust_root = try! IdentityKeyPair.generate()
-        let server_keys = try! IdentityKeyPair.generate()
-        let server_cert = try! ServerCertificate(keyId: 1, publicKey: server_keys.publicKey, trustRoot: trust_root.privateKey)
-        let sender_addr = try! SealedSenderAddress(e164: alice_address.name,
-                                                   uuidString: "9d0652a3-dcc3-4d11-975f-74d61598733f",
-                                                   deviceId: 1)
-        let sender_cert = try! SenderCertificate(sender: sender_addr,
-                                                 publicKey: alice_store.identityKeyPair(context: nil).publicKey,
-                                                 expiration: 31337,
-                                                 signerCertificate: server_cert,
-                                                 signerKey: server_keys.privateKey)
-
-        let message = Array("2020 vision".utf8)
-        let ciphertext = try sealedSenderEncrypt(message: message,
-                                                 for: bob_address,
-                                                 from: sender_cert,
-                                                 sessionStore: alice_store,
-                                                 identityStore: alice_store,
-                                                 context: nil)
-
-        let recipient_addr = try! SealedSenderAddress(e164: bob_address.name, uuidString: nil, deviceId: 1)
-        let plaintext = try sealedSenderDecrypt(message: ciphertext,
-                                                from: recipient_addr,
-                                                trustRoot: trust_root.publicKey,
-                                                timestamp: 31335,
-                                                sessionStore: bob_store,
-                                                identityStore: bob_store,
-                                                preKeyStore: bob_store,
-                                                signedPreKeyStore: bob_store,
-                                                context: nil)
-
-        XCTAssertEqual(plaintext.message, message)
-        XCTAssertEqual(plaintext.sender, sender_addr)
-    }
-
     static var allTests: [(String, (PublicAPITests) -> () throws -> Void)] {
         return [
             ("testAddreses", testAddress),
@@ -425,7 +278,6 @@ class PublicAPITests: XCTestCase {
             ("testHkdfSimple", testHkdfSimple),
             ("testHkdfUsingRFCExample", testHkdfUsingRFCExample),
             ("testGroupCipher", testGroupCipher),
-            ("testSessionCipher", testSessionCipher),
             ("testSenderCertifications", testSenderCertificates),
         ]
     }
